nvme/tcp: allow dissection from middle of NVMe/TCP session
When dissecting nvme traffic we need to know whether queue
is an admin queue or io queue as opcodes are interpreted differently.
According to the spec admin queue has an id of "0" and io queues > 0.
Unfortunately queue id information is only available during "fabrics connect"
command. Therefore if admin connect command was not captured there is no
way for us to obtain queue id.
However some opcode numbers can be only used by admin queue.
(for example io queue does not have opcode 0x18, [keepalive])
Therefore we can use this simple heuristics, i.e. if "fabrics connect" was not
captured, we consider every queue as io queue, but if we encounter opcode that
cannot be io opcode we switch to admin queue.